environments/dynastai/requirements.txt
   # filepath: /Users/torinvandenbulk/Documents/GitHub/atropos/environments/dynastai/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.0
python-multipart==0.0.6
uuid==1.30
aiohttp==3.8.5